Version 2.0.0 - Several changes...
Now "kopf.on.mutate" is used instead of "kopf.on.validate":
the k8s resource is directly changed to fulfill the requirements
(securityContext, gid), instead of checking that the user has
included that requirements.

Removed LOCAL_STORAGE in memory, not secure due to not persistent.
Now annotations on resource are used as recommended in doc:
https://kopf.readthedocs.io/en/latest/memos/#limitations

Avoid handle k8s garbage collector actions.

New annotation "chaimeleon.eu/test" is recognized if accessing datasets.
If it appears the request for access is sent to the test endpoint,
configurable with environment variable DATASET_SERVICE_TEST_ENDPOINT.
Note that the user GID is previously got from production endpoint
and put in the test endpoint to synchronize the DB.

Update base image to python:3.10
Update kopf version to 1.36.0

Better formatted sourcecode:
 - avoid long lines
 - avoid duplicated functions doing the same:
      now there is only one function with multiple decorators
 - other simplifications for better readability